Fixed merge errors
