+++ frontend/ya_delivery/src/App.js
@@ -2,11 +2,11 @@ import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import LoginForm from './Components/LoginForm/LoginForm'; 
 import RegistrationForm from './Components/RegistrationForm/RegistrationForm'; 
 import React from 'react';
-import GeolocationPage from '../LocationTest/Location';
-import Main from '../Main/Main';
-import ShopPage from '../ShopPage/ShopPage';
-import Cart from '../Cart/Cart';
-import AccountPage from '../AccountPage/AccountPage';
+//import GeolocationPage from './Components/LocationTest/Location';
+import Main from './Components/Main/Main';
+import ShopPage from './Components/ShopPage/ShopPage';
+import Cart from './Components/Cart/Cart';
+import AccountPage from './Components/AccountPage/AccountPage';
 
 import { ApolloClient, InMemoryCache, ApolloProvider, gql } from '@apollo/client';
 
@@ -27,7 +27,7 @@ const App = () => {
 
             <Route path="/" element= {<Main/>}/>  
             <Route path="main" element= {<Main/>}/>  
-            <Route path="geo" element= {<GeolocationPage/>}/>
+            {/* <Route path="geo" element= {<GeolocationPage/>}/> */}
 
 
             <Route path="shops/:id" element={<ShopPage/>}/>
