+++ frontend/ya_delivery/src/Components/AccountPage/AccountPage.jsx
@@ -1,3 +1,4 @@
+import React from 'react';
 
 const AccountPage = () => {
     return <>Здесь будет информация об аккаунте пользователя</>;
