+++ frontend/ya_delivery/src/Components/Assets/base-shop-image.png
