+++ frontend/ya_delivery/src/Components/Cart/Cart.jsx
@@ -1,3 +1,4 @@
+import React from 'react';
 
 const Cart = () => {
     return <>Здесь будет козрина с покупками</>
