+++ frontend/ya_delivery/src/Components/ShopPage/ShopPage.jsx
@@ -1,4 +1,5 @@
 import 'react-dom';
+import React from 'react';
 import { useParams } from 'react-router-dom';
 
 const ShopPage = (props) => {
