+++ frontend/ya_delivery/src/index.css
@@ -18,7 +18,7 @@ body {
   justify-content: center;
   align-items: center;
   min-height: 100vh;
-  background: url('./Components/Assets/pexels-dương-nhân-2232917.jpg') no-repeat;
+  /* background: url('./Components/Assets/pexels-dương-nhân-2232917.jpg') no-repeat; */
   background-size: cover;
   background-position: center;
 }
